OK-673: Siirretty henkilö-näkymän pistesyötön komponentit eri tiedostoihin
