+++ src/app/components/range.tsx
@@ -1,0 +1,9 @@
+import { NDASH } from '@/app/lib/constants';
+
+export const Range = ({
+  min,
+  max,
+}: {
+  min?: number | string;
+  max?: number | string;
+}) => (min || max ? `${min ?? ''}${NDASH}${max ?? ''}` : '');
